func_door: fix erratic behaviour of spawned trigger field
This was most noticeable with door .wait <= 1 but would eventually occur
at higher wait values if players hang around in the trigger field.

Also the door(s) usually did not stay open for the full .wait time after
the player left the trigger field, as some time had passed since the
.door_finished timestamp was set.

To make this reliable AND efficient requires updating the state of the
door(s), instead of maintaining a separate state in the trigger field entity.