func_door: update documentation to reflect default .dmg of 0
IMO 0 is the best default, because if it were > 0 then mappers would
have no way to configure a door that doesn't deal any damage when blocked.

The implemented default has been 0 since (at least) Nexuiz 2.5.2 so
changing it now would not be ideal for compatibility.

The comment about the default of 2 was inherited from Quake.